Fix compile errors on 64-bit platforms
git-svn-id: https://groke.otago.ac.nz/ant/ant@1138 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/index.c
@@ -133,7 +133,7 @@ readability->set_parser(parser);
 
 if (param_block.num_pregen_fields)
 	{
-	unsigned int pregen_prefix_len = strlen(param_block.index_filename) + 1;
+	size_t pregen_prefix_len = strlen(param_block.index_filename) + 1;
 	pregen = new ANT_pregen();
 
 	if (pregen_prefix_len < PATH_MAX)
